[LBEOS-32] net: Do not evaluate the nbr cache for lb_radio
THIS MUST NOT BE COMMITED UPSTREAM!!!

We should evaluate later if there is a better way to do this.
This commit fixes the situation where a packet is routed from ppp to
the lb_radio interface. When the routed packet is about to be send,
the nbr cache is evaluated to get the dst lladdr. This must not be done
for lb_radio!

Signed-off-by: Christian Taedcke <christian.taedcke@lemonbeat.com>