net: Fix ppp routing issue via peer addr
The function net_ipv6_prepare_for_send() implements determination of the
target network interface, the packet should be sent on. This is based on
the nbr list, routes and interface prefixes.

If a route is configuread via a ppp peer link local address, this
function did not behave correctly, because the target interface was just
copied from the pkt. But in case of multiple interfaces, the packet
might contain the wrong interface.

Now if the nexthop is not onlink (meaning not in the interface prefix),
the nbr list is checked to determine the correct target interface.

Signed-off-by: Christian Taedcke <christian.taedcke@lemonbeat.com>